docs/requirements.txt
myst-parser[sphinx,linkify]
pandas
pyyaml
python-hcl2
requests
sphinx-autobuild
sphinx-copybutton
sphinx-design
git+https://github.com/2i2c-org/sphinx-2i2c-theme
sphinxcontrib-mermaid
